+++ plugin-hrm-form/src/states/SearchContact.js
@@ -1,0 +1,40 @@
+import { HANDLE_SELECT_SEARCH_RESULT } from './ActionTypes';
+import searchResultMock from './SearchResultMock';
+
+export const handleSelectSearchResult = (searchResult, taskId) => ({
+  type: HANDLE_SELECT_SEARCH_RESULT,
+  searchResult,
+  taskId,
+});
+
+// TODO: implement real backend call
+export function searchContacts({ firstName, lastName, counselor, phoneNumber, dateFrom, dateTo }) {
+  return Promise.resolve(searchResultMock);
+}
+
+function copyNewValues(originalObject, objectWithNewValues) {
+  if (objectWithNewValues === null || typeof objectWithNewValues === 'undefined') {
+    return originalObject;
+  }
+
+  const isLeaf = originalObject.hasOwnProperty('value');
+
+  if (isLeaf) {
+    return {
+      ...originalObject,
+      value: objectWithNewValues,
+      touched: true,
+    };
+  }
+
+  const keys = Object.keys(originalObject);
+  const values = keys.map(key => copyNewValues(originalObject[key], objectWithNewValues[key]));
+
+  const entries = keys.map((key, i) => [key, values[i]]);
+  return Object.fromEntries(entries);
+}
+
+export function copySearchResultIntoTask(currentTask, searchResult) {
+  const searchResultDetails = searchResult.details;
+  return copyNewValues(currentTask, searchResultDetails);
+}
